+++ src/js/models/services.js
@@ -1,0 +1,11 @@
+// data structure to save the possible service options for the department
+export const Services = {
+    save : function(services){
+        localStorage.setItem('services-list', JSON.stringify(services));
+    },
+    list : function(){
+        return JSON.parse(localStorage.getItem('services-list')) || {};
+    }
+}
+
+export default Services;
