+++ src/js/views/04_personnel.js
@@ -3,7 +3,7 @@ import { View, ViewTable } from './view_class.js'
 import Table from "../components/table/table.js";
 import Form from "../components/form/form.js";
 
-import { Services } from "../utils/data_utils/budget_data_handlers.js";
+import Services from '../models/services.js';
 import FundLookupTable from '../models/fund_lookup_table.js';
 import { unformatCurrency } from "../utils/common_utils.js";
