+++ src/js/views/view_class.js
@@ -8,7 +8,8 @@ import Form from "../components/form/form.js";
 import Modal from "../components/modal/modal.js";
 
 import CurrentPage from "../models/current_page.js";
-import { AccountString } from "../utils/data_utils/budget_data_handlers.js";
+import AccountString from '../models/account_string.js'
+
 
 export class View {
